Updates aan de GUI
Nog wat andere updates aan de GUI zoals:
- Layout (ConstraintLayout)
- Nieuw Instellingen scherm aangemaakt
- Instellingenscherm geimplementeerd
