+++ SmokeS/app/src/main/AndroidManifest.xml
@@ -20,7 +20,8 @@
             </intent-filter>
         </activity>
         <activity android:name=".SelectTimeActivity" />
-        <activity android:name=".ProgressActivity"></activity>
+        <activity android:name=".ProgressActivity" />
+        <activity android:name=".SettingsActivity"></activity>
     </application>
 
 </manifest>
