+++ SmokeS/app/src/main/res/values/strings.xml
@@ -5,4 +5,9 @@
     <string name="navigation_drawer_close">Sluit navigatiebalk</string>
 
     <string name="action_settings">Instellingen</string>
+    <string name="geef_melding">Geef Melding</string>
+    <string name="geef_notificatie">Geef Notificatie</string>
+    <string name="hey_gebruiker">Hey \'Gebruiker\'!</string>
+    <string name="selecteer_tijd">Wijzig tijd</string>
+    <string name="update_test_data">update test data</string>
 </resources>
